refactor(settings): migrated repositories to generic side panel context
